Introduce new stage to filter the known false positive issues
